LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.BlobURLSharedWorker.sub.https.tentative.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".

Harness Error (TIMEOUT), message = null

TIMEOUT Verify a blob URL created via URL.createObjectURL from the third-party context shouldn't be useable as the shared worker script when it's passed to saa_handle.SharedWorker. Test timed out
